Remove "rad" prefix from newer routines in PsychRadiometric
I added this to a number of routines to minimize possibility of name space collisions.  But, it really clashes with the rest of the PTB, where this style is not adopted.  And, the routines in question had long and very likely to be unique names in any case.

This may break code of people who started using these routines.  But, I don't think we've actually had a release since I added these and in any case it should be easy to fix with a mutli-file search and replace.    In the long run we will be happier not to have the prefix.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>