Screen/GStreamer: Make movie bus message handling more robust.
If message processing is called, always iterate and handle all messages
until no more messages are pending on the bus. The doWait flag only
determines what should be done if there aren't any new messages pending,
ie., if call should return immediately or wait for up to 2 seconds.

Add an additional call to bus event processing in one of the while()
loops that
wait for new frames to arrive - in one of the retry loops.

Also a new routine to check if a movie is seekable, and some code
reformatting.
Additionally reduce verbosity level for rewind related status messages
from 6 to 5.