Screen/GStreamer: Fix passive frame fetching aka LoadMovieIntoTexturesDemo.
Make it work not only with other runtimes, but also with GStreamer SDK.
We follow the recommendation of basic GStreamer SDK tutorial 13 to only
send the seek event for frame stepping to the videosink, not to the
playbin2
itself. This seems to fix all the fetch failures, seek failures and
deadlocks we
experienced with our demo movie DualDiscs.mov or some other movies that
have sound tracks. Tested on OSX with SDK 2013.6.

Also improve error handling for step seeks and reduce timeout from 30
secs to
10 secs.

-> Screen for Matlab and Octave on OSX rebuilt and tested.