Screen: Fix wrong usage of OpenML defective flag.
Was checking for presence of kPsychOpenMLDefective flag in gfxcaps
variable instead of specialFlags in a few instances! This would output
a few wrong diagnostic messages in unlikely corner cases. It would also
choose the less precise / noise resistant timestamping path when our
homegrown frame-sequential stereo mode is in use on Linux + FOSS
drivers,
or during initial video refresh rate calibration. The latter is
problematic as
it will increase sensitivity to timing noise. This was detected when
testing
on Ubuntu 12.04.2 with an ancient RS480 gpu in an ancient laptop with
old
cpu and low ram memory. Bug went probably unnoticed due to its effect
mostly on very old hardware.

-> None of the possible symptoms is critical, so we'll just fold this
into
a future release.