Screen/GStreamer: Try (unsuccessfully) to make movie writing robust on Windows.
This tries to make movie file writing on Windows with the GStreamer SDK
robust.
Unfortunately it fails at it. The changes are supposed to potentially
increase robustness,
and as testing on OSX confirms, they don't hurt. However, the whole
thing doesn't
work any better on Windows (64-Bit Matlab with 2013.6 SDK). Writing a
valid movie
file curiously works, but Matlab hangs at the end of the writing
session and never
recovers.

Screen for OSX 64-Bit Matlab and octave as well as 64-Bit Matlab on
Windows rebuilt.
These changes are also included in the rebuilt Screen mex files from
previous commit.