Added GetGITInfo
This is a GIT version of GetSVNInfo.  It returns the result of
'git describe --always' which seems to be useful revision
information in all cases I've tested so far.  This only works
on directories, not on directories or files like GetSVNInfo.

The routine could probably be made more sophisticated.

I also added a GetGitPath to parallel GetSubversionPath.

In GetSVNInfo, noted in the help text that there is also a
path field in the returned struct.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>