Screen/OSX: Add new optional text layout method to ATSU as TextRenderer 0.
Screen('Preference','TextRenderer', method); now selects different
methods of text rendering with method == 0 vs. method == 1. Both
methods use Apple's OSX ATSU text renderer as in the past and the
default method 1 still behaves exactly as in the past. The optional
method = 0 setting now uses a different method of computing the text
bounding box and text baseline (for vertical positioning of text) than
before: It uses the deprecated ATSUMeasureTextImage() function for
computing the bounding box. Why? Because this way, the ATSU rendering
is able to cope with some unusual text fonts, e.g., Denis Pelli's
favorite "Kuenstler Script LT" font, which is misrendered with default
method 1 - It is clipped and mispositioned. The new method 0 also has
its own share of bugs, e.g., it mispositions and clips some other
"exotic" fonts like the japanese font we use in
DrawHighQualityUnicodeTextDemo. So now the user has the choice between
two evils ;) - depending on the font the user wants to use. Of course,
method 2, the Linux text rendering plugin, works perfectly fine and at
higher quality and performance with all tested standard and exotic
fonts, so switching to method 2 is the most simple and obvious choice.
But in the interest of consistency and backwards compatible we still
stick to the old ATSU method on OSX by default, so users who care about
every text pixel being perfectly at the same place as in the past will
not be taken by surprise.

-> ATSU text renderer code for OSX updated.
-> Screen for 64-Bit Matlab and Octave for OSX rebuilt and tested.
-> Some help text updates to DrawTextPlugin.m and
DrawHighQualityUnicodeTextDemo.m when we are here anyway.