iViewXComm: Add subfunction 'receivelast'.
Allows to fetch the most recent eyetracker data packet, e.g.,
if a queue of pending packets has built up, but we're interested
in live data.

See mesg #16300 and earlier for the reason. This solution is not
the greatest yet. Might be a good idea to switch the receive code
in this function to non-blocking operation (polling). Current blocking
operation can introduce a delay of ~0.5 msecs on Linux/OSX and ~1 msec
on Windows. This should be still ok'ish on Linux and OSX, but on Windows,
1 msec timeout will cause maximum sampling rate to drop below 1000 Hz,
which can be troublesome with the latest iView eyetrackers that can do
> 1000 Hz tracking.