Screen: NULL-out texture memory pointer received from GStreamer early.
NULL-out the reference to GStreamer owned and memory-managed video
memory buffer immediately after PsychCreateTexture() was used to create
an OpenGL texture from it. This is not really needed, as
PsychCreateTexture() does this already after texture creation if
clientstorage is not enabled, ie., always by default. We add this
explicit redundant NULL assignment as an annotation to the code, to
simplify correctness review of the code wrt. memory management somewhat.

-> Screen.mexw64 for 64-Bit Matlab on Windows rebuilt and tested.