Remove dead octave-3.0 backward compatibility code.
We don't need castDouble2Float() anymore on octave 3.2+. Delete it.
Therefore we don't need moglsingle() and mogldouble() anymore either,
but leave it here as a very thing 1-line wrapper unless we find time
to remove all calls to them. However, we must remove the dispatch()
optimization for octave, as the dispatch() function was deprecated
in octave 3.6 and is removed in 3.7+, so these functions would fail
on future stable octave releases.