Screen: Be more clever about 10 bpc framebuffers.
Check if a bpc=10 requested native framebuffer is set up 10 bpc via
query to OpenGL context. Only fallback to PsychtoolboxKernelDriver /
MMIO tricks, or eventually fail, if that is not the case.
This makes detection of need for MMIO fallback on Linux more clever.
This makes detection of need for PsychtoolboxKernelDriver fallback on
OSX more clever, as it at least tries to get 10 bpc from the OS, before
falling back to our own solution. Note that as of OSX 10.8, 10 bit
native framebuffers are still not supported, and likely this will also
be the case for the upcoming 10.9 Mavericks. This change in code is
just to make us behave intelligently if Apple ever decided to honor
requests by its customers.

Also fix handling of AGL conserveVRAM settings for OSX in 'OpenWindow'.