Screen/OSX: Drop support for AGL + Carbon, cleanup Cocoa path.
This removes all support for AGL + Carbon windows, and for OSX 10.5
Now we always use CGL for OpenGL, and for windowed mode or
half-transparent windows, we use CGL wrapped into NSOpenGL contexts +
windows created by Cocoa. This allowed for a nice cleanup of the
PsychCocoaGlue.c and PsychWindowGlue.c files.

-> Screen for 64-Bit Matlab and Octave on OSX rebuilt and tested.