Screen/OSX: Disable screen release/capture hack for cgl windows.
The hack was to release a captured display, then CGLClearDrawable to
detach the OpenGL context from the display, then capture the display
again. This was to fix some display corruption in stereo display mode
on OSX 10.5.x

As 10.5 is no longer supported, remove the workaround and just
CGLClearDrawable().
Fixes some warning messages in system log on 10.7.

-> Screen for OSX rebuilt and tested.