Update lib HIDAPI for Linux with latest bug fixes:
Merge upstream's master as of 10th September 2013 in order to
get the setlocale() fix for Linux which caused a very ugly bug, as
reported by Ian Mackenzie in forum message #16555.

This bug changed the LC_ALL, and thereby LC_NUMERIC locale setting
of the running octave process to system default, from "C". That
in turn caused decimal points "." no longer to be recognized by
octave and thereby parsing failure for all decimal numbers in scripts
or at the command line: Throwing away everything after the decimal
point caused all numbers to be floor()'ed to integers!

HIDAPI fixed this problem: It only changes the LC_CTYPE locale instead
of LC_ALL, so the LC_NUMERIC which is critical to us won't be touched
anymore in any case. Additionally it only changes LC_CTYPE if the
application hasn't set it itself already, thereby avoiding to mess
with app settings.

-> Renamed hid-libusb.c to hid.c for consistency across platforms.
-> Will require a rebuilt of PsychHID for Linux.