GetMouse, mogldrawdots3D: Micro-Optimizations for speed.
GetMouse: Make sure, lazy evaluation in if-statement evaluates
the cheaper condition first, to shove of a few microseconds.

moglDrawDots3D: ismember() is apparently a rather expensive function
in both octave and matlab, so only use it for checking against large
target sets, not for trivial checks for only 2 values, as we did.
This should save significant time in parameter validation and thereby
function execution.