Screen: Improve drawtext plugins for Linux, add font caching.
The drawtext rendering plugins on Linux (and optionally OSX) now
maintain separate sets of OGLT or FTGLES font objects for use in
OpenGL text rendering for each onscreen window. For this purpose,
Screen() now passes the numeric parent onscreen window handle for
each drawingtarget (onscreen/offscreen/texture) as a "contextId"
to the drawtext plugins. The plugins distinguish different windows
and thereby associated OpenGL contexts and maintain settings for
each of them.

This is to allow that OpenGL resources like display lists and glyph
textures or glyph atlas textures are not shared between master OpenGL
contexts, but stored separately. This makes text rendering work if
OpenGL context resource sharing is either disabled by usercode, or
impossible, e.g., because different onscreen windows are located
on different X-Screens and therefore can't share resources.

-> DrawText now works if multiple onscreen windows are opened on
   different X-Screens. Before it malfunctioned for all but the
   first opened window due to the unsatisfiably OpenGL context
   resource sharing requirement.

Additionally, a LRU cache is now implemented for font objects
corresponding to different font settings of a context. The cache can
hold a certain amount of font objects and performs LRU replacement
on each context, if the cache gets full. This way, switching between
different text settings on Linux no longer slows down the succeeding
drawtext call if the font objects are already in the cache.

-> Provides constant speed text rendering when switching text settings,
   wheras before, a slowdown of about 10x - 15x was observed after a
   setting change. Typical numbers from a 2010 MacBookPro with NVidia
   GeForce 330M: About ~0.3 msecs per "Hello World!" text draw with
   constant or switched settings. Before the caching, switching settings
   caused a slowdown to about 3.5 msecs for first draw after switch.

-> Screen() for 64-Bit Octave on Linux rebuilt with new plugin api context
   support and tested.

-> libptbdrawtext_ftgl[es]64.so.1 for 64-Bit Linux with OpenGL / OpenGL-ES
   modified, rebuilt and tested.

A new benchmark DrawTextFontSwitchSpeedTest.m allows to benchmark performance
of text switching on a system.

Additionally the text renderer on Linux now returns the true font family name
aka Screen('TextFont') of a selected font, *after* a Screen('DrawText') call,
so scripts can actually check if they got the font they wanted, or just a
substitute font, because the requested font wasn't installed on the system.

Some robustness improvements were made to FontInfo.m to test and demonstrate
this.

-> All Linux and OSX Screen / drawtextplugin mex files need rebuild before release.