Add Super-Joana's SuperShapeDemo :)
A new OpenGL demo, demonstrating rendering of so called Supershapes.
Drawing / rotating the shape is fast. Creation/Updating of the shapes
shape is relatively slow, due to computation on the CPU. This could
only get much faster by use of GLSL Vertex- and Geometryshading on the
GPU.

This demo was contributed by Joana Leitao.

-> PsychDemos/OpenGL4MatlabDemos/SuperShapeDemo.m added.