DaqAInScan: Add the hack used by user masondar for 64-Bit OSX.
64-Bit OSX has problems detecting all Daq interfaces properly.
Add some hack that worked for one user on the forum "masondar123",
e.g., ptb forum message thread from June 7 2013.

If we are on 64-Bit OSX, aka IsOSX(1), we turn detection error()'s
into pure warning()'s and hard-code the 'IndexRange' of interfaces
to scan, instead of using the flawed detection results.

Testing on my USB-1408FS showed that DaqAInScan works reliably
on Linux and Windows, and sort of ok on 32-Bit OSX. On 64-Bit OSX
this hack doesn't help me on 10.5.8 Leopard or 10.7.x Lion, but
masondar reported success on a 10.8 system with USB-1608FS. If this
is due to different devices, or different OS'es, or different
runtimes or just luck, who knows? Anyway, add this hack to make
stuff work at least on the one setup (masondar's) were it apparently
can work. Hopefully this also helps others...

-> DaqAInScan.m modified.