Fix bug in ComputeNakaRushton.  This was not a pretty bug.  Not sure when it was introduced.
The code was dividing the contrast in the numerator by sigma before exponentiation.  This made no sense and is not what the comments said the routine did.  I fixed it to match the comments (which correctly describe the Naka-Rushton.)

If there is code somewhere that counts on this bug, that code will now break.  But it probably should rather than trying to preserve backwards compatibility with what this routine was doing.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>