Screen: Allow use of multisample textures for multisample anti-aliasing.
If GL_ARB_texture_multisample extension is supported, then preferrably
create and use multisample textures as color attachments for the
imaging pipeline drawBufferFBO's if multisampling is requested. Only
fallback to good ol' multisampled renderbuffers if textures fail or are
unsupported.

Multisample textures have the advantage that they can be used with the
imaging pipelines identity blit chain, and thereby by the panelfitter
fallback code. This should allow to provide framebuffer rotation for
multisampled framebuffers via the panelfitter.

It also provides yet another fallback for multisample resolve if scaled
multisample resolves are not supported.

So far the theory. In practice the code-path is only compile-tested,
because Apple's crappy os doesn't expose the needed extension under
OpenGL 2.1, despite the gpu being perfectly capable of supporting it.
-> Will need reboot to a real os for testing.