BitsPlusImagingPipelineTest: Make more lenient for acceptable errors.
Tolerance for small conversion errors can now be set via optional
'acceptableerror' parameter. It defaults to 1 unit of output luminance
or color intensity, ie., an allowable conversion error of up to 1 part
in 65535 or an error in the least significant bit of the 16 bit output
signal.

Improve plotting, diagnostic, help text etc.

-> Lenience for a 1 unit error allows Intel HD (Gen 5 "Ironlake"
family) gpu's to pass the test without detrimental effects on actual
experiments. This makes this card useable for such purpose in principle.