Screen/Linux: Always poll-wait in PsychOSGetSwapCompletionTimestamp()
Not always of course, but also poll-wait if there is only 1 onscreen
window and all other requirements are fulfilled. We need to avoid a
blocking wait for swapbuffer completion, as we hold the display lock
during such a wait, so we'd block not only the (non-existent in the
single-window case) other graphics threads, but also the main thread
when executing X11 calls, e.g., GetMouse() queries.

This wasn't an issue before, but now that we lock the display lock, it
could cause some delays on the main thread even in the single window
case.