Reenable PsychCV for 64-Bit Octave on Linux.
This provides access to the ARToolkit tracking functions on
64-Bit GNU/Octave on Linux. As ARToolkit is GPLv2 licensed,
it - and therefore our PsychCV.mex file linked against it -
can only be used under GNU/Octave which is GPL licensed as
well. We only enable this on 64-Bit Octave for Linux, as a
little bonus for Linux users, if at all.

This just for testing. May get disabled before next beta release
again.

-> PsychCV for 64-Bit Octave on Linux built and tested.
-> ARToolkitDemo made functional again with GStreamer video capture
   engine as default.