Screen: Fix small bug in PsychNormalizeTextureOrientation().
Didn't restore the framebuffer binding to zero, as it should,
because PsychSetDrawingTarget(NULL); as used there would be
a no-op mostly for the purpose of resetting the binding to zero.

Fixed. Fixes a "invalid operation" error on nouveau. However, this
was not the culprit of morphing errors in MorphDemo. These remain
a mystery.