Screen/OSX: Rewrite screen glue to remove/fix use of deprecated functions.
Rewrite screen glue to use functions which are supported/current as of
OSX 10.9, to replace functions which are deprecated since 10.6.
Add missing function prototype of CGDisplayBeamPosition() - which was
completely removed in the OSX 10.9 SDK - until we decide if we should
already kill it completely, or let it linger around for a while longer.
Current status:
10.6 - Only works on NVidia gpus, 10.7 - Ditto, but slightly
broken/inaccurate - PTB driver is better solution, 10.8 Severely broken
for everything bug VGA-CRT's, 10.9 completely dysfunctional (always
returns zero on NVidia card).
Remove unused PsychGetNumScreenPlanes() function for all os'es as a
cleanup.

-> Screen for 64-Bit Matlab on OSX rebuilt and lightly tested.