Screen/OSX: Replace ATSUI text renderer by CoreText text renderer.
ATSUI is deprecated since 10.6 and difficult to support on 10.9+, so
let it die.
Screen('DrawText') for OSX specific textrenderer (type 0 and 1)
reimplemented by use of CoreText framework. Additional code cleanup.
As a little side effect, the CoreText renderer now supports
Screen('TextModes') which never had any function in PTB-3 and it
supports "underlined" text style - in theory, as i couldn't find any
font which actually supports that. As another bonus, the renderer can
now handle the "Kuenstler LT" script font which was so painful to use
in the old implementation.

-> Screen and FontInfo for 64-Bit Octave and Matlab on OSX rebuilt and
tested.