Screen/OSX: Optimize font switching a tiny bit more.
Cache the CTFontDescriptorRef in our font structs and select the font
in Screen('DrawText') directly from the font descriptor, instead of
going through postscript names. The speedup is insignificant though, in
the low microsecond range.

I also checked were most of the time during font enumeration is spent.
It is not in opening the CoreText font, but in querying font metrics
like ascenders, descenders etc. These queries apparently trigger an
iteration over all glyphs in the font file to compute those properties
and that takes up over 96% of the processing time, accounting for about
14 seconds of the 14-15 seconds startup time in FontInfo(). Anyway,
these metrics queries are mostly unavoidable if FontInfo() is supposed
to be useful for much, so we'll have to live with this.

-> FontInfo and Screen for 64-Bit Matlab and Octave on OSX rebuilt and
tested.