Screen/OSX: Add AppNap protection et al. for OSX 10.9+
Whenever Screen() gets initialized, AppNap protection is enabled.
Whenever it gets clear'ed, AppNap protection is disabled. As Screen()
is also called by Priority() on OSX, this will imply protection for
anything visual and anything using Priority() to imply it is timing
sensitive.

What does it do? On pre 10.9 nothing. On 10.9+ it uses the new
"Activity" API to tell the OS to prevent or enable certain
actions/optimizations to make sure the system doesn't go to sleep or
nap, or the display doesn't go to sleep or nap, or the application
isn't starved of cpu cycles (=forced to nap) during PTB operations. It
also tells the OS that this app is "NSActivityLatencyCritical".

We somewhat abuse th api's here, which will lead to shortened battery
life as long as Screen() isn't flushed, but correctness is more
important than battery life, and i don't have time or nerves to think
of a more intelligent and balanced approach. Better safe than sorry.

-> Screen for 64-Bit Matlab and Octave on OSX rebuilt and tested.
-> While it seems to work and doesn't seem to cause problems i also
can't see any measureable benefit in timing behaviour. On my machine
OSX 10.9 just sucks display-timing wise and i haven't tested other
modalities.