Screen: Honor a new optional specialflags setting 'kPsychGUIWindowWMPositioned'
If specified together with kPsychGUIWindow, it tells Screen() to leave
window placement of onscreen GUI windows to the window manager instead
of to Screen.
This will ignore the [left, top, ...] parameter in the 'rect'. Size of
the window will still be defined by the 'rect', but not its absolute
position. The hope is that the window manager will do a better job at
window placement, although testing on OSX 10.9 Mavericks shows that at
least on OSX this hope was not warranted.

-> Screen for 64-Bit Octave and Matlab on OSX rebuilt and tested.
-> Screen on other os'es pending.
-> New flag definition file kPsychGUIWindowWMPositioned.m added.