Screen: Allow lossless writing of movies with 1, 3 or 4 layers and 8 or 16 bpc.
The Screen(‘CreateMovie’) function now optionally accepts a target
number of color channels (1 = Red/Grayscale, 3 = RGB, 4 = RGBA) instead
of using a fixed RGBA layout. It also accepts an optional bitdepth
argument to allow for 8 bpc or 16 bpc encoding, instead of the
previously fixed 8 bpc. Screen(‘AddFrameToMovie’) can now readback the
framebuffer as Red or RGB in addition to the old RGBA, and readback 16
bpc data in addition to the old 8 bpc. Exception is embedded OpenGL-ES
graphics, where only 8 bpc RGB or RGBA is supported, but not 16 bpc or
Red only.

-> The movie writing functions strive to writeout data in the requested
format and layout. This makes 8/16 bpc writeout and pure grayscale or
RGB writeout possible if suitable codecs and muxers are selected for
encoding. Optimizations are in place for the ffenc_huffyuv and
ffenc_ljpeg encoders: These are lossless codecs, so we try to feed them
data in a format that actually allows lossless encoding, at least for 8
bpc color resolution. Only certain color caps and color channel layouts
allow this.

-> These codecs are supposed to be lossless, but verification of this
hypothesis is pending.
-> These codecs only support 8 bpc, so 16 bpc so far leads to loss of
precision down to 8 bpc. Still hunting for a 16 bpc capable
encoding/codec that actually works.
-> The huffyuv codec works well for en/decoding and playback, but
produces relatively large files, with only ~ 2x compression for RGB and
almost no compression for Grayscale. The ljpeg codec provides higher
compressions of 2x - 4x, but GStreamer can’t decode ljpeg data, so this
is problematic for ptb use. vlc can decode such movies, so apparently
encoding works, but we’d need a transcoding step to use such footage in
ptb.

-> Screen for 64-Bit Matlab on OSX rebuilt and tested.