Screen: Add > 8 bpc video capture support to GStreamer video capture engine.
Well, sort of. The code now honors the ‘bitdepth’ parameter and tries
to switch to 16 bpc video capture and/or recording if a bitdepth > 8
bpc is requested. How well this actually works depends a lot on the
operating system and type of GStreamer installation and set of plugins:
- Video recording > 8 bpc untested and not likely to work at the moment.
- Video capture with camerabin1 has a chance of working with supported
cameras and suitable video capture plugins, but so far only osx tested,
which doesn’t support > 8 bpc at the moment, so this was more a
simulation of 16 bpc gray which looks promising. Tests on Linux and
Windows are pending.
- Video capture with camerabin2 probably can’t achieve > 8 bpc because
camerabin2 requires camerawrapperbinsrc to operate and that plugin has
a hard-coded videocrop filter element in its processing path and
videocrop filter as of GStreamer-0.10 only supports 8 bpc. Things might
be much better with GStreamer-1.0 but for the moment we are still stuck
on the v0.10 series due to other limitations of the 1.0 series.
-> This is more a forward/future looking enhancement, with hopefully
pays off once we port to GStreamer 1.0 in the future.

-> Screen for 64-Bit Matlab on OSX rebuilt and tested.