Screen: Add "stop of capture" framecount barrier to dc1394 capture.
We need to make sure that when synchronized stop of capture on
multi-cam synced setups is requested, that all recording threads
dequeue and push the same number of frames from the engine before
terminating, so we don't end up with different final framecounts
in the video files between different cameras.

This adds an implementation which uses the framecount at end of capture
of the master as reference value for all threads to stop.

The target stop count can also get manually queried and set, as
well as the current and possible capture counts.

-> Screen for 64-Bit Octave on Linux rebuilt and tested.
-> Demo updated for testing.
-> Lightly tested with bus sync and seems to work there. More tests tbd.