PsychHID/Linux: Fix "XFCE-4 creates new workspaces at KbQueueStop" problem.
XFCE-4 for whatever reason created a new work space each time
a call to KbQueueStop caused emission (XSendEvent()) of a
keypress event with keycode 0 (in order to unblock the KbQueueThread
to facilitate its fast termination).

Now we do the same trick, but send a key release event instead of
a key press event, and magically the problem goes away!

-> PsychHID for 64-Bit Octave on Linux rebuilt and successfully
   tested on XFCE-4 desktop environment on 64-Bit Ubuntu 12.04.2 LTS.