Screen: Add support for offline debayering of recorded raw video.
A new optional Screen('OpenMovie', ...) specialFlags1 & 1024 setting
tells the playback engine that the video footage is actually raw
sensor data which needs a Bayer filtering step to convert it into
RGB color images. If the flag is set accordingly and pixelFormat=1
is chosen, then libdc1394's debayering code is used for debayering
during playback. The Screen('SetVideoCaptureParameter', ) function
allows choice of debayer method and bayer sensor pattern for playback
if the capturehandle -1 is specified. Default settings are the same
as for video capture, ie., nearest-neighbour and pattern 0.

Update help texts to describe new features of dc1394 capture engine
and movie playback.

-> Screen for 64-Bit Octave on Linux rebuilt and tested.
-> VideoMultiCameraCaptureDemo.m contains new test code to test
   Bayer vs. normal recording, lossless huffyuv and ffenc_sgi codecs,
   and high bit depths 16 bpc capture and recording. Also some other
   comments have been cleaned up.

=> This should conclude more or less the first multicam capture/recording
   implementation via libdc1394, assuming all pre-release tests go well.