osxmakeitoctave: Adapt to Octave's expected version roadmap.
Octave 3.6.4 was the last release of the Octave 3.6 series.
It will be followed by Octave 3.8.0 very soon, and possibly a 3.8.1 or
- much less likely - a 3.8.2. Therefore cut off our support at 3.6.4
and add 3.8.0 - 3.8.2 as successors.
We also add (optimistically) support for the next-gen 4.0 series with
4.0.0 and 4.0.1, in the unlikely case that the naming schemes and path
structures will stay exactly the same in the future. Who even knows if
our 3.8.x path scheme will be valid?