PsychtoolboxKernelDriver for 64-Bit OSX rebuilt.
To add the latest PCI Id’s for AMD GPU’s as of end of 2013.

Also needed to update the Info.plist to require other versions of
kernel kpi’s, so no link/load failure happens on f%$##% OSX 10.9
Mavericks. Let’s hope this doesn’t break support for the kernels of
10.6 - 10.8.

-> New PsychtoolboxKernelDriver64Bit.kext.zip file.
-> No update for dead 32-Bit driver.