Update Nvidia GPU handling for NV-E0 and later.
NV-E0 “Kepler” and later cards need different registers to be read for
retrieving the VTOTAL and vblbias values needed for beam position
timestamping correction. So much is clear from Ben Suppnik’s nouveau
patches. Therefore add new handling code for the NV-E0+ case. What
isn’t clear yet is if the register base addresses in Ben’s patch are
correct - due to lack of testing hardware.

Another open question is if NV-D0 GPU’s belong into the C0 or the E0
category? For now we assign them to C0, in accordance with the nouveau
Wiki’s Code->Names mapping page at
http://nouveau.freedesktop.org/wiki/CodeNames/
However, this contradicts the nouveau-kms patches which treat D0 the
same as E0 and F0? Which one is right?

-> Code applies to Linux and OSX.

-> Screen for 64-Bit Octave 3.8 on OSX rebuilt and tested on NV-A0
type, NV-50 class Nvidia GeForce GT 330M.

-> Testing of NV-50 code path shows surprisingly that the VACTIVE reg
(according to nouveau ams) contains the same vtotal values as the
VTOTAL register according to PTB. The SYNC_START_TO_BLANK_END reg of
PTB contains the same value as VBLANKE from nouveau-kms, which is
expected and reassuring. This means that the nv-d0 code of Ben is
probably correct and applicable for us as well.