Screen/dc1394: Add new flag for latency reduction in multi-cam capture.
A new optional 'SyncMode' flag 32 (kPsychNoLockStepSync) allows to
minimize processing latency in the video capture/processing/recording
threads of the dc1394 engine. Setting the flag will cause the slave
camera threads to no longer wait for the master camera thread to
reach a framecount, ie., all threads will work in parallel and the
slaves won't wait for the master to finish processing a framecount
before starting their processing themselves. This avoids adding a
delay of the duration of processing on the master thread.

-> Helpful for low-latency video capture / video processing on
   multi-cam setups.

-> Downside: While cams still start and execute capture in sync,
   they may stop asynchronously, ie., some cameras may capture or
   record one or multiple more frames than the master thread due
   to variability in scheduling on the os and the firewire bus.
   This is usually not a problem for capture and tracking apps,
   but could be a slight hassle for video recording apps.

-> Screen for 64-Bit Octave on Linux rebuilt and tested.