VideoDVCamCaptureDemo: Performance tuning for MS-Windows.
Testing on MS-Windows (Windows-7) with both 32-Bit and 64-Bit Matlab
R2010b and a Sony DV-PAL camera shows that the code as shipped in the
30-Dec-2013 beta works. However, use of video/x-dv caps is not
necessary, and performance is suboptimal with the dvdec decoder. Remove
superfluous caps and switch to ffdec_dvvideo, which shows good
performance - capture at 25 fps, as expected.

Testing with theoraenc video codec, but *no sound* shows that pure
video recording with theoraenc works well at full performance and
resolution as well.

-> All good in Windows-land at lest for PAL-DV firewire cameras, no
testing with NTSC-DV possible.
-> Btw. while dshowvideosrc works well, ksvideosrc doesn’t work at all
with either DV or DC1394 cameras. We should make sure to prefer
dshowvideosrc over ksvideosrc, as it seems to be much more robust with
many cameras.