PsychImaging('EnableNative10BitFramebuffer'): Execute LoadIdentityClut() during setup.
This is not needed for identity LUT loading, as the hardware LUT’s are
bypassed in 10 npc mode anyway, but as a side-effect it also disables
output dithering on all outputs. This was not necessary in the good’ol
days of analog VGA CRT monitors, but now that 10 bit digital flat
panels are a reality, we do need to explicitly disable output dithering.
-> Reported by Nicolas Cottaris when testing Radeon HD 4870 with a Sony
PVM-2541A 10 bit panel.
-> This fixes ops on that panel and gives well working 10 bit output,
as tested on OSX 10.8.5 with its
PsychtoolboxKernelDriver64BitLegacy.kext.