PsychKinect: Update docu and help text for "Microsoft Kinect for Windows"
Turns out libfreenect version 0.2 (for Linux) and 0.3 (for OSX) and
later do now support the Kinect for Windows boxes, not only the
original XBOX-360 kinect. Explain this to user.
Our MS-Windows driver is still XBOX-360 Kinect only, as it doesn’t use
libfreenect as backend, but a special MS-Windows version of a Kinect
backend which is only api compatible to libfreenect, but not abi
compatible. That backend doesn’t do “Kinect for Windows” yet.