Screen/OSX: Reimplement conserveVRAM setting 8192.
kPsychUseAGLForFullscreen now behaves closer to its original behaviour
before the switch to a pure Cocoa windowing backend. It still uses
Cocoa for fullscreen window setup, but still performs target display
capture (aka CGDisplayCapture()), raising the Cocoa window to the
window level of the shielding window, so rendering works but other apps
are locked out of a display.

The difference to kPsychUseAGLCompositorForFullscreen aka 16384 is that
that setting doesn’t capture the display, so composited operation is
possible.

None of these settings helps in any way to fix the brokeness on OSX
10.9.x with Nvidia.
However there are some hints that 16384 (and possibly 8192) may help to
fix the dual-window failures on OSX 10.8.5 with AMD and possibly other
GPU’s.

-> Add OSXCompositorIdiocyTest.m as a test script for checking such
stuff.