Native10Bit frame buffer: Do not disable digital display dithering by default.
Now LoadIdentityClut() allows to specify if digital display dithering
should be disabled or not, as part of identity LUT setup. By default,
dithering gets disabled, as in the past, to retain backwards
compatibility.
If the optional 4th parameter ‘disableDithering’ is set to zero, then
the dither disable step is skipped and dithering control is left to the
OS and graphics driver.

PsychImaging() still disabled dithering by default for any display mode
that also needs a high precision output and an identity clut. This to
keep things like Bits+, Datapixx etc. working as in the past.

PsychImaging(‘AddTask’, ‘General’, ‘EnableNative10BitFramebuffer’)
however does still ask for identity gamma tables to provide 10 npc
frame buffer images untampered to the encoders, but it no longer
disables dithering on digital outputs anymore by default. The decision
to dither or not to dither is left to the OS and graphics driver. The
optional ‘disableDithering’ parameter can be set to 1 to force
dithering off. The idea here is that dithering can have some value,
e.g., as another psycho-visual method to simulate 10 bpc stimuli on 8
bpc displays, similar to our ‘PseudoGray’ bit stealing implementation.
A spec compliant graphics driver should decide based on display EDID
data if dithering is needed or not and do the right thing automatically.

-> Update PsychImaging, also clarify current status of 10 bpc support
in the help text, with what we’ve learned during the last days.
-> LoadIdentityClut updated.
-> Screen updated to allow setup of identity passthrough without dither
disable.

-> Screen for 64-Bit Octave on OSX rebuilt, but only lightly tested,
due to lack for suitable hardware.
-> We can only automatically influence dithering on AMD hardware, and i
only have Nvidia hw here.