Screen/OSX: Keep compatibility to deprecated, outdated kernel driver revision 0 for now.
We are still nominally supporting OSX 10.6, and 10.6 ships on some
hardware with a 32-Bit kernel, so the 32-bit kernel driver is required.
We can’t update the 32-Bit driver anymore due to lack of SDK and
compilers, so that driver is eternally frozen to revision 0. Keep
Screen workable on rev 0 drivers. However, if a AMD GPU with a pre
DCE-4 display engine is detected (Radeon HD-4000 an earlier), output a
warning about the need to upgrade to revision 1 driver if any kind of
high bit depth stimulus presentation is to be done, as the old rev 0
driver only works correctly on single-display setups with HD-3000 /
DCE-2 and earlier, and doesn’t disable dithering at all on HD-4000 /
DCE-3. This can cause less robust operation on devices like Bits+ /
Bits# / Data/View/ProPixx, 10 bit frame buffers and digital panels.

-> Screen for 64-Bit Octave and Matlab on OSX rebuilt.